v0.3.2-alpha edits
Changes:

It is now possible to drag in order to remove or place obstacles.

Fixed generation numbering.

Improved performance on larger levels (improved draw and collision functions).

NOTE: old levels must be opened with the GDE level editor, "fixed" (by pressing the 'X' key) and saved in order for the old level to work in newer versions.